fix some adpater logic
